backend/requirements.txt
# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Data Generation
faker==22.0.0
bcrypt==4.1.2